+++ backend/src/main/java/cn/zl/account/book/controller/response/AccountInfoResponse.java
@@ -26,11 +26,11 @@ public class AccountInfoResponse {
     @JsonSerialize(using= ToStringSerializer.class)
     private Long accountOwnershipId;
 
-    private Long accountBalance;
+    private Double accountBalance;
 
-    private Long accountIncome;
+    private Double accountIncome;
 
-    private Long accountExpenditure;
+    private Double accountExpenditure;
 
     private Integer accountType;
 }
